fixed problem with pidgin music tracker
git-svn-id: svn://svn.code.sf.net/p/qmmp-dev/code/trunk@1886 90c681e8-e032-0410-971d-27865f9a5e38

+++ qmmp/src/plugins/General/mpris/tracklistobject.cpp
@@ -90,11 +90,11 @@ QVariantMap TrackListObject::GetMetadata(int in0)
         map.insert("artist", item->artist());
         map.insert("album", item->album());
         map.insert("tracknumber", item->track());
-        map.insert("time", item->length());
-        map.insert("mtime", item->length() * 1000);
+        map.insert("time", (quint32)item->length());
+        map.insert("mtime", (quint32)item->length() * 1000);
         map.insert("genre", item->genre());
         map.insert("comment", item->comment());
-        map.insert("year", item->year());
+        map.insert("year", item->year().toUInt());
     }
     return map;
 }
